# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Warung Buah API",
        "description": "API Documentation untuk aplikasi Warung Buah - Sistem manajemen transaksi buah dengan fitur lengkap untuk supplier, produk, kategori, transaksi, dan pembayaran.",
        "contact": {
            "name": "API Support",
            "email": "support@warungbuah.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/transaksi/{id}": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Detail transaksi",
                "description": "Dapatkan detail transaksi beserta item, peti, biaya, dan pembayaran",
                "operationId": "getTransaksiDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Transaksi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TransaksiDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Hapus transaksi",
                "description": "Hapus transaksi beserta semua data terkait",
                "operationId": "deleteTransaksi",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Transaksi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaksi berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaksi berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transaksi/form-data": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Data untuk form transaksi",
                "description": "Dapatkan data yang dibutuhkan untuk form transaksi (komisi default, supplier, produk)",
                "operationId": "getTransaksiFormData",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "komisi_default": {
                                                    "type": "number",
                                                    "example": 10
                                                },
                                                "suppliers": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Supplier"
                                                    }
                                                },
                                                "stok_tersedia": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/DetailBarangDatang"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transaksi/statistics": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Statistik transaksi",
                "description": "Dapatkan ringkasan statistik transaksi",
                "operationId": "getTransaksiStatistics",
                "parameters": [
                    {
                        "name": "tanggal_dari",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "tanggal_sampai",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total_transaksi": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "total_pendapatan": {
                                                    "type": "number",
                                                    "example": 15000000
                                                },
                                                "total_piutang": {
                                                    "type": "number",
                                                    "example": 5000000
                                                },
                                                "transaksi_lunas": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "transaksi_tempo": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "transaksi_cicil": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TransaksiDetail": {
                "title": "Transaksi Detail",
                "description": "Model Transaksi dengan relasi",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Transaksi"
                    },
                    {
                        "properties": {
                            "user": {
                                "$ref": "#/components/schemas/User"
                            },
                            "item_transaksi": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ItemTransaksi"
                                }
                            },
                            "biaya_operasional": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/BiayaOperasional"
                                }
                            },
                            "pembayaran": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Pembayaran"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "BiayaOperasional": {
                "title": "Biaya Operasional",
                "description": "Model Biaya Operasional",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "transaksi_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_biaya": {
                        "type": "string",
                        "example": "Ongkos kirim"
                    },
                    "nominal": {
                        "type": "number",
                        "example": 50000
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token yang didapat dari endpoint login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}